Rename decimal.h to pydecimal.h.
